+++ src/core/ids/network.ts
@@ -1,3 +1,5 @@
-export const POLYGON = 'ReFDxuVQ674gHd2kFjsFL';
-export const ETHEREUM = 'JnWfsCw2gqoPWFT1NXmFWW';
-export const GEO = 'KJjKetFsGVSbw9qFpRzRSy';
+import { Id } from '../../id.js';
+
+export const POLYGON = Id('ReFDxuVQ674gHd2kFjsFL');
+export const ETHEREUM = Id('JnWfsCw2gqoPWFT1NXmFWW');
+export const GEO = Id('KJjKetFsGVSbw9qFpRzRSy');