+++ src/core/image.ts
@@ -5,14 +5,22 @@
  * @since 0.0.6
  */
 
-import { generate } from '../id.js';
+import { type Id, generate } from '../id.js';
 import { Relation } from '../relation.js';
 import { SystemIds } from '../system-ids.js';
-import type { CreateRelationOp, SetTripleOp } from '../types.js';
+import type { Op } from '../types.js';
 
 type MakeImageReturnType = {
-  imageId: string;
-  ops: [CreateRelationOp, SetTripleOp];
+  id: Id;
+  ops: Array<Op>;
+};
+
+type MakeImageParams = {
+  cid: string;
+  dimensions?: {
+    width: number;
+    height: number;
+  };
 };
 
 /**
@@ -20,36 +28,62 @@ type MakeImageReturnType = {
  *
  * @example
  * ```ts
- * const { imageId, ops } = Image.make('https://example.com/image.png');
- * console.log(imageId); // 'gw9uTVTnJdhtczyuzBkL3X'
+ * const { id, ops } = Image.make({ cid: 'https://example.com/image.png', dimensions: { width: 100, height: 100 } });
+ * console.log(id); // 'gw9uTVTnJdhtczyuzBkL3X'
  * console.log(ops); // [...]
  * ```
  *
- * @returns imageId – base58 encoded v4 uuid representing the image entity: {@link MakeImageReturnType}
+ * @returns id – base58 encoded v4 uuid representing the image entity: {@link MakeImageReturnType}
  * @returns ops – The ops for the Image entity: {@link MakeImageReturnType}
  */
-export function make(src: string): MakeImageReturnType {
+export function make({ cid, dimensions }: MakeImageParams): MakeImageReturnType {
   const entityId = generate();
-
-  return {
-    imageId: entityId,
-    ops: [
-      Relation.make({
-        fromId: entityId,
-        toId: SystemIds.IMAGE_TYPE,
-        relationTypeId: SystemIds.TYPES_PROPERTY,
-      }),
-      {
-        type: 'SET_TRIPLE',
-        triple: {
-          entity: entityId,
-          attribute: SystemIds.IMAGE_URL_PROPERTY,
-          value: {
-            type: 'URL',
-            value: src,
-          },
+  const ops: Array<Op> = [
+    Relation.make({
+      fromId: entityId,
+      toId: SystemIds.IMAGE_TYPE,
+      relationTypeId: SystemIds.TYPES_PROPERTY,
+    }),
+    {
+      type: 'SET_TRIPLE',
+      triple: {
+        entity: entityId,
+        attribute: SystemIds.IMAGE_URL_PROPERTY,
+        value: {
+          type: 'URL',
+          value: cid,
         },
       },
-    ],
+    },
+  ];
+
+  if (dimensions) {
+    ops.push({
+      type: 'SET_TRIPLE',
+      triple: {
+        entity: entityId,
+        attribute: SystemIds.IMAGE_HEIGHT_PROPERTY,
+        value: {
+          type: 'NUMBER',
+          value: dimensions.height.toString(),
+        },
+      },
+    });
+    ops.push({
+      type: 'SET_TRIPLE',
+      triple: {
+        entity: entityId,
+        attribute: SystemIds.IMAGE_WIDTH_PROPERTY,
+        value: {
+          type: 'NUMBER',
+          value: dimensions.width.toString(),
+        },
+      },
+    });
+  }
+
+  return {
+    id: entityId,
+    ops,
   };
 }