+++ src/graph/index.ts
@@ -1,5 +1,5 @@
 export * from './create-entity.js';
+export * from './create-image.js';
 export * from './create-property.js';
 export * from './create-space.js';
 export * from './create-type.js';
-export * from './publish.js';